+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Unreleased]
+
+### Fixes
+- Fixed pod status scheduled race condition between the scheduler and the pod binding
+
+
 ## [v0.5.1] - 2025-05-20
 
 ### Added